+++ src/jk_cmdoutputparsinghelper/__init__.py
@@ -1,13 +1,19 @@
-﻿
 
 
-__version__ = "0.2021.12.30"
+
+__author__ = "Jürgen Knauth"
+__version__ = "0.2023.2.9"
 
 
 
 from ._LineUtils import getPositionsOfWords
 
+from .TextSpanInfo import TextSpanInfo
+from .TextSpanInfoSequence import TextSpanInfoSequence
+from .SpaceInfoSequence import SpaceInfoSequence
+
 from .LineList import LineList
+from .CharMatrix import CharMatrix, ColumnPositionInfo
 from .ColumnDef import ColumnDef
 from .Table import Table
 from .TextData import TextData
@@ -15,3 +21,5 @@ from .TextData import TextData
 from .ValueParserDef import ValueParserDef
 from .LineParser_ParseAtFirstDelimiter import LineParser_ParseAtFirstDelimiter
 from .ValueParser_ByteWithUnit import ValueParser_ByteWithUnit
+
+